+++ DlssUpdater/changelog.md
@@ -1,6 +1,7 @@
 ﻿# 1.0.5.0
 * Add support for hiding games from main list with right mouse button
 * Add filter to games page
+* Validate MD5 hash and signature of downloaded DLL
 
 # 1.0.4.1
 * Fix bug with image on manual added game which would not remove text